requirements.txt
# Core Dependencies
pdfplumber==0.10.3 
pandas==2.1.4
openpyxl==3.1.2
streamlit==1.29.0

# General Utilities
Pillow==10.1.0
python-dateutil==2.8.2
pytz==2023.3
PyPDF2==3.0.1
python-dotenv==1.0.0

# Gemini API Dependency (REPLACED anthropic)
google-genai 

# Other Dependencies
flask==3.0.0
werkzeug==3.0.1
gunicorn==21.2.0